+++ NEWS.md
@@ -1,3 +1,8 @@
+## Version 2.6 (2020-11-22)
+
+- The cache system is now session-persistent with `cache.dir = file.path(tempdir(), 'BGS_Cache')`. This solves the problem with mismatching price series from cached data between splits or dividends. A new warning is set whenever the user uses cache.dir different from temp.dir()
+- removed dplyr grouping message
+
 ## Version 2.5.9 (2020-11-17)
 
 - removed rownames from output ([issue 18](https://github.com/msperlin/BatchGetSymbols/issues/18))